+++ src/Application/Events.Application.Services/Features/Users/Commands/Delete/DeleteUserHandler.cs
@@ -1,17 +1,19 @@
-﻿using Events.Application.Services.Features.Users.Repositories;
-using Events.Application.Services.Features.Users.Specifications;
+﻿using Events.Application.Services.Shared;
+using Events.Domain.Aggregates.Users;
+using Events.Domain.Aggregates.Users.Errors;
+using Events.Domain.Exceptions;
 using MediatR;
 
 namespace Events.Application.Services.Features.Users.Commands.Delete;
 
-/// <inheritdoc />
-public sealed class DeleteUserHandler(IUserRepository userRepository) : IRequestHandler<DeleteUserCommand>
+public sealed class DeleteUserHandler(IRepository<User> userRepository) : IRequestHandler<DeleteUserCommand>
 {
-    /// <inheritdoc />
     public async Task Handle(DeleteUserCommand request, CancellationToken cancellationToken)
     {
-        var spec = new UserSpec().WithId(request.Id);
-        var user = await userRepository.GetAsync(spec, cancellationToken);
+        var user = await userRepository.GetByIdAsync(request.Id, cancellationToken);
+
+        if (user == null)
+            throw new NotFoundException(UserErrorMessages.UserNotFoundById(request.Id));
 
         await userRepository.DeleteAsync(user, cancellationToken);
     }
